backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
jobspy
httpx
numpy
faiss-cpu
sqlite-utils